refactor(frontend): decompose feed state, prompt voice input, and draft button
